+++ lib/src/concurrency/droppable/debounce_droppable_handler.dart
@@ -10,8 +10,7 @@ import 'package:streamed_controller/src/concurrency/droppable/droppable_handler.
 /// Additional handler over [DroppableConcurrency]
 /// Forces a controller to wait a [debounceDuration] before process stream
 /// Resets that [Timer] if user interupted timer with any event
-class DebouncedDroppableConcurrencyHandler<State extends Object>
-    extends DroppableConcurrencyHandler<State> {
+class DebouncedDroppableConcurrencyHandler<State extends Object> extends DroppableConcurrencyHandler<State> {
   Timer? _$debounceTimer;
 
   final Duration debounceDuration;
@@ -21,13 +20,14 @@ class DebouncedDroppableConcurrencyHandler<State extends Object>
             'Use DroppableConcurrencyHandler instead if you\'re doesn\'t wants to apply debouncing to event handler');
 
   @override
-  Future<void> handle(
-      Stream<State> $stream, void Function(State) stateCallback) async {
+  Future<void> handle(Stream<State> $stream, void Function(State) stateCallback) async {
     // For debouncing handle calls
     _$debounceTimer?.cancel();
-    _$debounceTimer =
-        Timer(debounceDuration, () => super.handle($stream, stateCallback));
+    _$debounceTimer = Timer(debounceDuration, () => super.handle($stream, stateCallback));
 
     return;
   }
+
+  @override
+  String toString() => 'Droppable, debounce: $debounceDuration';
 }