+++ lib/src/concurrency/sequental_handler.dart
@@ -3,8 +3,7 @@ import 'dart:async';
 import 'package:streamed_controller/src/concurrency/base_handler.dart';
 import 'package:streamed_controller/src/utils/async_queue.dart';
 
-class SequentalConcurrentHandler<State extends Object>
-    extends HandlerBase<State> {
+class SequentalConcurrentHandler<State extends Object> extends HandlerBase<State> {
   int _processingCalls = 0;
 
   final _$queue = AsyncQueue<void>();
@@ -18,14 +17,10 @@ class SequentalConcurrentHandler<State extends Object>
   }
 
   @override
-  Future<void> handle(
-      Stream<State> $stream, void Function(State) stateCallback) async {
+  Future<void> handle(Stream<State> $stream, void Function(State) stateCallback) async {
     Future<void> $queueObject() {
       final $subscription = $stream.listen(stateCallback);
-      var $future = $subscription
-          .asFuture<void>()
-          .whenComplete(() => _cancelSub($subscription))
-          .onError((e, s) {
+      var $future = $subscription.asFuture<void>().whenComplete(() => _cancelSub($subscription)).onError((e, s) {
         _cancelSub($subscription);
         if (e != null) {
           Error.throwWithStackTrace(e, s);
@@ -49,5 +44,7 @@ class SequentalConcurrentHandler<State extends Object>
     _$queue.close();
   }
 
+  @override
+  String toString() => 'Sequental';
   SequentalConcurrentHandler({this.eventTimeout});
 }