+++ lib/src/observer/controller_observer.dart
@@ -10,8 +10,8 @@ class StreamedControllerObserver {
 
   factory StreamedControllerObserver.dartLog() => _TestControllerObserver();
 
-  static Never _overrideError() => throw UnimplementedError(
-      'This method should be overrided with your own logging package/solution');
+  static Never _overrideError() =>
+      throw UnimplementedError('This method should be overrided with your own logging package/solution');
 
   @mustBeOverridden
   void onCreate(StreamedController controller) => _overrideError();
@@ -20,31 +20,29 @@ class StreamedControllerObserver {
   void onDispose(StreamedController controller) => _overrideError();
 
   @mustBeOverridden
-  void onStateChanged(
-          StreamedController controller, Object prevState, Object nextState) =>
-      _overrideError();
+  void onStateChanged(StreamedController controller, Object prevState, Object nextState) => _overrideError();
 
   @mustBeOverridden
-  void onError(StreamedController? controller, Object? error,
-          StackTrace stackTrace) =>
-      _overrideError();
+  void onError(StreamedController? controller, Object? error, StackTrace stackTrace) => _overrideError();
+
+  @mustBeOverridden
+  void onEvent(StreamedController? controller, String eventName) => _overrideError();
 }
 
 class _TestControllerObserver extends StreamedControllerObserver {
   @override
-  void onCreate(StreamedController<Object> controller) =>
-      log('${controller.runtimeType} created');
+  void onCreate(StreamedController<Object> controller) => log('[$controller] Controller created');
 
   @override
-  void onDispose(StreamedController<Object> controller) =>
-      log('${controller.runtimeType} disposed');
+  void onDispose(StreamedController<Object> controller) => log('[$controller] Controller disposed');
   @override
-  void onError(StreamedController<Object>? controller, Object? error,
-          StackTrace stackTrace) =>
-      log('${controller.runtimeType} got error: $error');
+  void onError(StreamedController<Object>? controller, Object? error, StackTrace stackTrace) =>
+      log('[$controller] Error: $error');
 
   @override
-  void onStateChanged(StreamedController<Object> controller, Object prevState,
-          Object nextState) =>
-      log('${controller.runtimeType} state updated, new state: $nextState');
+  void onStateChanged(StreamedController<Object> controller, Object prevState, Object nextState) =>
+      log('[$controller] State updated: $nextState');
+
+  @override
+  void onEvent(StreamedController<Object>? controller, String eventName) => log('[$controller] Event: $eventName');
 }